backend/requirements.txt
# Django ve Rest Framework
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=3.7.0
django-filter>=21.1
channels>=3.0.0
drf-yasg>=1.20.0

# Veritabanı
psycopg2-binary>=2.9,<3.0
dj-database-url>=0.5.0

# API ve Veri Kaynakları
yfinance>=0.2.3
requests>=2.25.1
python-dotenv>=0.19.0
pandas>=1.3.0
numpy>=1.21.0
websocket-client>=1.2.1

# ML ve Veri Analizi
scikit-learn>=1.0.0
tensorflow>=2.8.0
transformers>=4.15.0
nltk>=3.6.5
textblob>=0.15.3
vaderSentiment>=3.3.2
joblib>=1.1.0
prophet>=1.1.3
pystan>=3.5.0
statsmodels>=0.13.5

# Google Cloud
google-cloud-language>=2.6.0

# Görselleştirme
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Yardımcı Araçlar
tqdm>=4.62.0
schedule>=1.1.0
Pillow>=8.3.0
pyyaml>=6.0
celery>=5.2.0
redis>=4.0.0
jwt>=1.3.0

# Dokümantasyon
Sphinx>=4.3.0